cursor pointer on icons?
Hi Sean, I am just testing to see your opinion on adding a class to the icons that show the whois popup and the the remove icon that will just add the cursor:pointer style to the icons.  At first I was thinking about the bootstrap button style, but I'm not sure how that would look with multiple rows of the buttons together.  I added a proposed change to the jailDisplay.css also.  (also added the class to the remove icon for the fail regexes too?)