Apache TOTP google authentication
This will setup apache authentication with TOTP Google Authentication.
I believe this can be 2 factor authentication based upon how you create the users ".google_authenticator" files.  I believe you have to add "PASSWORD=SECRET" to the file for the 2nd password.  Then the authentication parameters are "username" "SECRET12341" if the totp value at the time is 12341.  If the "PASSWORD=SECRET" is not added to the .google_authenticator file for the user, then the password would just be "12341".